Add architecture documentation To match the documentation, rename interface to presentation
