+++ src/app.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { UserModule } from './interface/user/user.module';
+import { UserModule } from './presentation/rest/user/user.module';
 import {PostgresModule} from "./infrastructure/postgres/postgres.module";
 
 @Module({
