+++ src/main.ts
@@ -10,7 +10,6 @@ async function bootstrap() {
       .setTitle('Wiredcraft assignment')
       .setDescription('Test backend for user, friends and geolocation')
       .setVersion('1.0')
-      // .addTag('cats')
       .build();
   const document = SwaggerModule.createDocument(app, config);
   SwaggerModule.setup('api', app, document);
